Initial work on #91
