+++ tests/features/overview.feature
@@ -65,7 +65,6 @@ Feature: Overview
 
     Given I am logged in as a user with the "access patterns page" permission
     And I am on "/patterns/media"
-
     Then the response should contain "/ui_patterns_test_theme/templates/patterns/media/css/media1.css"
     And the response should contain "/ui_patterns_test_theme/templates/patterns/media/css/media2.css"
     And the response should contain "/ui_patterns_test_theme/templates/patterns/media/js/media1.js"
