+++ tests/src/Kernel/UiPatternBaseTest.php
@@ -5,6 +5,7 @@ namespace Drupal\Tests\ui_patterns\Kernel;
 use function bovigo\assert\assert;
 use function bovigo\assert\predicate\isTrue;
 use function bovigo\assert\predicate\equals;
+use Drupal\Component\Serialization\Yaml;
 use Drupal\ui_patterns\UiPatternBase;
 
 /**
@@ -34,8 +35,7 @@ class UiPatternBaseTest extends AbstractUiPatternsTest {
 
     foreach ($definitions as $definition) {
       /** @var \Drupal\ui_patterns\UiPatternBase $pattern */
-      $arguments = [[], 'plugin_id', $definition];
-      $pattern = $this->getMockForAbstractClass(UiPatternBase::class, $arguments);
+      $pattern = $this->getUiPatternBaseMock($definition);
 
       assert($pattern->getId(), equals($definition['id']));
       assert($pattern->getLabel(), equals($definition['label']));
@@ -53,6 +53,41 @@ class UiPatternBaseTest extends AbstractUiPatternsTest {
         assert($pattern->getFieldPreview($name), equals($definition['fields'][$name]['preview']));
       }
     }
+  }
+
+  /**
+   * Test hookLibraryInfoBuild.
+   *
+   * @covers ::hookLibraryInfoBuild
+   */
+  public function testHookLibraryInfoBuild() {
+    $items = Yaml::decode(file_get_contents(dirname(__FILE__) . '/../fixtures/libraries.yml'));
+
+    foreach ($items as $item) {
+      $pattern = $this->getUiPatternBaseMock($item['actual']);
+
+      /** @var \Drupal\ui_patterns\UiPatternBase $pattern */
+      $libraries = $pattern->getLibraryDefinitions();
+      assert($libraries, equals($item['expected']));
+    }
+  }
+
+  /**
+   * Get UiPatternBase mock.
+   *
+   * @param array $configuration
+   *    Plugin configuration.
+   * @param array $methods
+   *    List of methods to mock.
+   *
+   * @return \PHPUnit_Framework_MockObject_MockObject
+   *    Mock object.
+   */
+  protected function getUiPatternBaseMock(array $configuration = [], array $methods = []) {
+    $root = \Drupal::service('app.root');
+    $typed_data_manager = \Drupal::service('typed_data_manager');
+    $arguments = [[], 'plugin_id', $configuration, $root, $typed_data_manager];
+    return $this->getMockForAbstractClass(UiPatternBase::class, $arguments, '', TRUE, TRUE, TRUE, $methods);
   }
 
 }
